Correct mistake when detecting candidate letters
I have made some adjustments to the findlegallyconnectedcomponents(rays). The if conditions when looking to the ratios should be && instead of || operators. Since you will always have a true value with ||, which is not the purpose of that statement. Now the ratio should be bounded to 3.0 like described in the algorithm.